feat(ecommerce): add support for products batch id request

+++ projects/ecommerce/src/app/shared/local-storage.service.ts
@@ -1,22 +1,31 @@
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Injectable, PLATFORM_ID, inject } from '@angular/core';
 
 /**
  * localStorage wrapper. Browser-only.
  * Should not be used during the server-side
  * rendering phase (no use case).
  */
-// We intentionally don't check if the platform is browser.
-// Otherwise, we won't get an error on the server, if we
-// happen to use LS during the render phase. As a result,
-// any problems caused by that might go unnoticed or hard
-// to track.
 @Injectable({ providedIn: 'root' })
 export class LocalStorage {
+  private _platformId = inject(PLATFORM_ID);
+
+  private get _isBrowser() {
+    return isPlatformBrowser(this._platformId);
+  }
+
   get(key: string): string | null {
+    if (!this._isBrowser) {
+      return null;
+    }
     return localStorage.getItem(key);
   }
 
   getJSON(key: string): object | null {
+    if (!this._isBrowser) {
+      return null;
+    }
+
     const item = this.get(key);
 
     try {
@@ -27,18 +36,34 @@ export class LocalStorage {
   }
 
   set(key: string, data: string) {
+    if (!this._isBrowser) {
+      return;
+    }
+
     localStorage.setItem(key, data);
   }
 
   setJSON(key: string, data: object) {
+    if (!this._isBrowser) {
+      return;
+    }
+
     localStorage.setItem(key, JSON.stringify(data));
   }
 
   remove(key: string) {
+    if (!this._isBrowser) {
+      return;
+    }
+
     localStorage.removeItem(key);
   }
 
   clear() {
+    if (!this._isBrowser) {
+      return;
+    }
+
     localStorage.clear();
   }
 }
